+++ trunk/mpf-rest-api/src/main/java/org/mitre/mpf/rest/api/pipelines/transients/TransientPipelineDefinition.java
@@ -25,45 +25,48 @@
  ******************************************************************************/
 
 
-package org.mitre.mpf.rest.api.pipelines.temp;
+package org.mitre.mpf.rest.api.pipelines.transients;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.google.common.collect.ImmutableList;
-import org.hibernate.validator.constraints.NotBlank;
-import org.mitre.mpf.rest.api.pipelines.ActionProperty;
+import org.hibernate.validator.constraints.NotEmpty;
+import org.mitre.mpf.rest.api.util.AllNotBlank;
 import org.mitre.mpf.rest.api.util.Utils;
 
 import javax.validation.Valid;
-import java.util.Collection;
+import java.util.List;
 
-public class TransientAction {
+public class TransientPipelineDefinition {
 
-    private final String _name;
-    @NotBlank
-    public String getName() {
-        return _name;
-    }
-
-    private final String _algorithm;
-    @NotBlank
-    public String getAlgorithm() {
-        return _algorithm;
-    }
-
-    private final ImmutableList<ActionProperty> _properties;
+    private final ImmutableList<String> _pipeline;
+    @NotEmpty
     @Valid
-    public ImmutableList<ActionProperty> getProperties() {
-        return _properties;
+    public ImmutableList<@AllNotBlank String> getPipeline() {
+        return _pipeline;
     }
 
-    public TransientAction(
-            @JsonProperty("name") String name,
-            @JsonProperty("algorithm") String algorithm,
-            @JsonProperty("properties") Collection<ActionProperty> properties) {
-        _name = Utils.trimAndUpper(name);
-        _algorithm = Utils.trimAndUpper(algorithm);
-        _properties = properties == null
+    private final ImmutableList<TransientTask> _tasks;
+    @Valid
+    public ImmutableList<TransientTask> getTasks() {
+        return _tasks;
+    }
+
+    private final ImmutableList<TransientAction> _actions;
+    @Valid
+    public ImmutableList<TransientAction> getActions() {
+        return _actions;
+    }
+
+    public TransientPipelineDefinition(
+            @JsonProperty("pipeline") List<String> pipeline,
+            @JsonProperty("tasks") List<TransientTask> tasks,
+            @JsonProperty("actions") List<TransientAction> actions) {
+        _pipeline = Utils.trimAndUpper(pipeline, ImmutableList.toImmutableList());
+        _tasks = tasks == null
                 ? ImmutableList.of()
-                : ImmutableList.copyOf(properties);
+                : ImmutableList.copyOf(tasks);
+        _actions = actions == null
+                ? ImmutableList.of()
+                : ImmutableList.copyOf(actions);
     }
 }
