+++ trunk/workflow-manager/src/main/java/org/mitre/mpf/wfm/service/pipeline/PipelineService.java
@@ -32,7 +32,7 @@ import org.mitre.mpf.rest.api.pipelines.Action;
 import org.mitre.mpf.rest.api.pipelines.Algorithm;
 import org.mitre.mpf.rest.api.pipelines.Pipeline;
 import org.mitre.mpf.rest.api.pipelines.Task;
-import org.mitre.mpf.rest.api.pipelines.temp.TransientPipelineDefinition;
+import org.mitre.mpf.rest.api.pipelines.transients.TransientPipelineDefinition;
 import org.mitre.mpf.wfm.data.entities.persistent.JobPipelineElements;
 
 public interface PipelineService {
