+++ trunk/workflow-manager/src/main/java/org/mitre/mpf/wfm/service/pipeline/PipelineValidator.java
@@ -30,9 +30,9 @@ package org.mitre.mpf.wfm.service.pipeline;
 import com.google.common.collect.ImmutableMultiset;
 import com.google.common.collect.Multiset;
 import org.mitre.mpf.rest.api.pipelines.*;
-import org.mitre.mpf.rest.api.pipelines.temp.TransientAction;
-import org.mitre.mpf.rest.api.pipelines.temp.TransientPipelineDefinition;
-import org.mitre.mpf.rest.api.pipelines.temp.TransientTask;
+import org.mitre.mpf.rest.api.pipelines.transients.TransientAction;
+import org.mitre.mpf.rest.api.pipelines.transients.TransientPipelineDefinition;
+import org.mitre.mpf.rest.api.pipelines.transients.TransientTask;
 import org.mitre.mpf.rest.api.util.Utils;
 import org.mitre.mpf.wfm.service.WorkflowPropertyService;
 import org.springframework.stereotype.Service;
